backend/cicd-integrity-monitor-main/scanner/requirements.txt
pyyaml
rich
jinja2
gitpython
jinja2
toml
GitPython
requests
